Added a default SSLFactory for the client
